Fix when active project/toDo items get updated
Before, the activeID would preemptively update to undefined when one
of the addItem buttons is pressed. Now, even if the addItem buttons are
pressed, the activeId only updates if the submit button is pressed. If
the user closes the modal without submitting, the activeID stays the
same as before, and the DOM does not update.